delete review route fixed

+++ app/api/reviews_routes.py
@@ -1,19 +1,21 @@
 from flask import Blueprint, jsonify, request, url_for, render_template, redirect, flash
 from flask_login import login_required, current_user
-from app.models import  db, User, Business, Review
+from app.models import db, User, Business, Review
 from app.forms import CreateReviewForm
 from werkzeug.utils import secure_filename
 import os
 
 review_routes = Blueprint('reviews', __name__)
 
-#get all reviews
+# get all reviews
+
+
 @review_routes.route('/')
 def get_all_reviews():
     all_reviews = Review.query.all()
     return {"reviews": [reviews.to_dict() for reviews in all_reviews]}
 
-#get review by id
+# get review by id
 # @review_routes.route('/<business_id/rev>/')
 # def get_review_id(business_id):
 #     reviews = Review.query.get(Review.business_id == business_id).all()
@@ -23,7 +25,7 @@ def get_all_reviews():
 #     # return review.to_dict()
 
 
-#create a review
+# create a review
 @review_routes.route('/<business_id>/reviews', methods=['POST'])
 @login_required
 def create_review(business_id):
@@ -32,16 +34,18 @@ def create_review(business_id):
 
     if form.validate_on_submit():
         new_review = Review(
-            review = form.data['review'],
-            rating = form.data['rating'],
-            user_id = form.data['user_id'],
-            business_id = form.data['business_id']
+            review=form.data['review'],
+            rating=form.data['rating'],
+            user_id=form.data['user_id'],
+            business_id=form.data['business_id']
         )
         db.session.add(new_review)
         db.session.commit()
         return new_review.to_dict()
 
-#update a review
+# update a review
+
+
 @review_routes.route('/<int:id>', methods=['PUT'])
 @login_required
 def update_review(id):
@@ -58,15 +62,19 @@ def update_review(id):
     else:
         return "404: Unauthorized User"
 
-#delete a review
-@review_routes.route('/<review_id>', methods=['DELETE'])
+# delete a review
+
+
+@review_routes.route('/<int:review_id>', methods=['DELETE'])
 @login_required
 def delete_review(review_id):
     review = Review.query.get(review_id)
-    print(review, 'reeeeeeeeeview')
-    if current_user.id == review.user_id:
-        db.session.delete(review)
-        db.session.commit()
-        return "Review has been successfully deleted"
+    if review:
+        if current_user.id == review.user_id:
+            db.session.delete(review)
+            db.session.commit()
+            return "Review has been successfully deleted"
+        else:
+            return "404 Unauthorized User"
     else:
-        return "404 Unauthorized User"
+        return "Review not found"
